Mapeo de entidades JPA, repositorios con JPQL y data.sql configurados

+++ demo/src/main/java/com/example/demo/entities/Espacio.java
@@ -14,11 +14,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Entity
 public class Espacio {
-    // Representa un lugar que se puede reservar dentro del sistema.
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 120)
     private String nombre;
